Move the PanTrackEvent code into its own files.
