+++ MusicLib/Types/Track/TrackEvents/channelVoiceEvents.hpp
@@ -16,52 +16,6 @@
 
 namespace bw_music {
 
-    struct MUSICLIB_API PanTrackEvent : public TrackEvent {
-        DOWNCASTABLE(PanTrackEvent, TrackEvent);
-        STREAM_EVENT(PanTrackEvent);
-
-        /// Construct from a value in the range [0,.. 0x80000000,.. 0xFFFFFFFF].
-        static PanTrackEvent fromValue32(ModelDuration timeSinceLastEvent, std::uint32_t highResValue) {
-            return PanTrackEvent(timeSinceLastEvent, highResValue);
-        }
-
-        /// Construct from a value in the range [0,.. 2^(numSourceBits - 1),.. (2^numSourceBits) - 1].
-        template<std::uint8_t numSourceBits>
-        static PanTrackEvent fromValue(ModelDuration timeSinceLastEvent, std::uint32_t value) {
-            return PanTrackEvent(timeSinceLastEvent, bw_music::minCentreMaxScaleUp<numSourceBits, 32>(value));
-        }
-
-        /// Construct from a value in the range [-1.0,.. 0.0,.. 1.0].
-        static PanTrackEvent fromSignedNormalizedDouble(ModelDuration timeSinceLastEvent, double signedNormalizedValue) {
-            return PanTrackEvent(timeSinceLastEvent, bw_music::assertScaleSignedNormalizedDoubleto32(signedNormalizedValue));
-        }
-
-        /// Get a value in the range [0,.. 0x80000000,.. 0xFFFFFFFF].
-        std::uint32_t getValue32() const;
-
-        /// Construct from a value in the range [0,.. 2^(numSourceBits - 1),.. (2^numSourceBits) - 1].
-        template<std::uint8_t numSourceBits>
-        std::uint32_t getValue() const {
-            return bw_music::minCentreMaxScaleDown<32, numSourceBits>(m_value);
-        }
-
-        /// Get a value in the range [-1.0,.. 0.0,.. 1.0].
-        double getSignedNormalizedValue() const {
-            return bw_music::scale32toSignedNormalizedDouble(m_value);
-        }
-
-        std::size_t getHash() const override;
-
-      protected:
-        /// Construct from a value in the range [0,.. 0x80000000,.. 0xFFFFFFFF].
-        PanTrackEvent(ModelDuration timeSinceLastEvent, std::uint32_t highResValue);
-
-        bool doIsEqualTo(const TrackEvent& other) const override;
-
-      private:
-        std::uint32_t m_value;
-    };
-
     struct MUSICLIB_API VolumeTrackEvent : public TrackEvent {
         DOWNCASTABLE(VolumeTrackEvent, TrackEvent);
         STREAM_EVENT(VolumeTrackEvent);
